refactor(api): migrate to feature-based structure
