+++ apps/api/src/features/cars/index.ts
@@ -1,0 +1,4 @@
+export * from "./queries";
+export { carsRoutes } from "./routes";
+export * from "./schemas";
+export * from "./service";
