+++ apps/api/src/features/coe/index.ts
@@ -1,0 +1,3 @@
+export * from "./queries";
+export { coeRoutes } from "./routes";
+export * from "./schemas";
