+++ apps/api/src/features/health/index.ts
@@ -1,0 +1,2 @@
+export { healthRoutes } from "./routes";
+export * from "./schemas";
