+++ apps/api/src/features/health/schemas.ts
