+++ apps/api/src/features/months/index.ts
@@ -1,0 +1,2 @@
+export { monthsRoutes } from "./routes";
+export * from "./schemas";
