+++ apps/api/src/lib/workflows/posts.ts
@@ -1,6 +1,6 @@
+import { getCarsAggregatedByMonth } from "@api/features/cars/queries";
+import { getCoeForMonth } from "@api/features/coe/queries";
 import { generatePost } from "@api/lib/gemini/generate-post";
-import { getCarsAggregatedByMonth } from "@api/queries";
-import { getCoeForMonth } from "@api/queries/coe";
 import { tokeniser } from "@api/utils/tokeniser";
 import type { WorkflowContext } from "@upstash/workflow";
