--- apps/api/src/queries/index.ts
@@ -1,2 +1,0 @@
-export * from "./cars";
-export * from "./coe";
